wip: Move types
