+++ docs/data/toolpad/core/components/sign-in-page/CredentialsSignInPage.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { AppProvider } from '@toolpad/core/AppProvider';
-import { SignInPage, type AuthProvider } from '@toolpad/core/SignInPage';
+import { SignInPage } from '@toolpad/core/SignInPage';
+import type { AuthProvider } from '@toolpad/core/auth';
 import { useTheme } from '@mui/material/styles';
 
 // preview-start
