+++ docs/data/toolpad/core/integrations/nextjs-approuter.md
@@ -181,7 +181,8 @@ Use the `SignInPage` component to add a sign-in page to your app. For example, `
 
 ```tsx title="app/auth/signin/page.tsx"
 import * as React from 'react';
-import { SignInPage, type AuthProvider } from '@toolpad/core/SignInPage';
+import { SignInPage } from '@toolpad/core/SignInPage';
+import type { AuthProvider } from '@toolpad/core/auth';
 import { AuthError } from 'next-auth';
 import { providerMap, signIn } from '../../../auth';
