+++ docs/src/components/landing/ToolpadAuthDemo.tsx
@@ -3,7 +3,8 @@ import Paper from '@mui/material/Paper';
 import { HighlightedCode } from '@mui/docs/HighlightedCode';
 import DemoSandbox from 'docs/src/modules/components/DemoSandbox';
 import { AppProvider } from '@toolpad/core/AppProvider';
-import { SignInPage, type AuthProvider } from '@toolpad/core/SignInPage';
+import { SignInPage } from '@toolpad/core/SignInPage';
+import type { AuthProvider } from '@toolpad/core/auth';
 import Frame from '../../modules/components/Frame';
 
 const NOOP = () => {};
