+++ examples/core/auth-nextjs-pages-nextauth-4/src/pages/auth/signin.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import type { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import Link from '@mui/material/Link';
-import { SignInPage, type AuthProvider } from '@toolpad/core/SignInPage';
+import { SignInPage } from '@toolpad/core/SignInPage';
+import type { AuthProvider } from '@toolpad/core/auth';
 import { getProviders, signIn } from 'next-auth/react';
 import { getServerSession } from 'next-auth/next';
 import { useRouter } from 'next/router';
