+++ examples/core/auth-nextjs/src/app/auth/signin/page.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { SignInPage, type AuthProvider } from '@toolpad/core/SignInPage';
+import { SignInPage } from '@toolpad/core/SignInPage';
+import type { AuthProvider } from '@toolpad/core/auth';
 import { AuthError } from 'next-auth';
 import { providerMap, signIn } from '../../../auth';
