+++ packages/create-toolpad-app/src/templates/auth/auth.ts
@@ -1,4 +1,4 @@
-import type { SupportedAuthProvider } from '@toolpad/core/SignInPage';
+import type { SupportedAuthProvider } from '@toolpad/core/auth';
 import { kebabToConstant, kebabToPascal } from '@toolpad/utils/strings';
 import { requiresIssuer, requiresTenantId } from './utils';
 import { Template } from '../../types';
