+++ packages/create-toolpad-app/src/templates/auth/utils.ts
@@ -1,4 +1,4 @@
-import type { SupportedAuthProvider } from '@toolpad/core/SignInPage';
+import type { SupportedAuthProvider } from '@toolpad/core/auth';
 
 export function requiresIssuer(provider: SupportedAuthProvider) {
   return (
